refactor(ts): separate wc session requests into files
